requirements.txt
beautifulsoup4==4.10.0; python_full_version > "3.0.0"
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
charset-normalizer==2.0.7; python_full_version >= "3.6.0" and python_version >= "3"
click==8.0.3; python_version >= "3.6"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.6" and python_full_version >= "3.5.0"
flask-caching==1.10.1; python_version >= "3.5"
flask==2.0.2; python_version >= "3.6"
gitdb==4.0.9; python_version >= "3.7"
gitpython==3.1.24; python_version >= "3.7"
gunicorn==20.1.0; python_version >= "3.5"
idna==3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
importlib-metadata==4.8.1; python_version < "3.8" and python_version >= "3.6"
itsdangerous==2.0.1; python_version >= "3.6"
jinja2==3.0.2; python_version >= "3.6"
markupsafe==2.0.1; python_version >= "3.6"
requests==2.26.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
smmap==5.0.0; python_version >= "3.7"
soupsieve==2.2.1; python_version >= "3.6" and python_full_version > "3.0.0"
typing-extensions==3.10.0.2; python_version < "3.8" and python_version >= "3.7"
urllib3==1.26.7; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
werkzeug==2.0.2; python_version >= "3.6"
zipp==3.6.0; python_version < "3.8" and python_version >= "3.6"
